pyproject.toml
[project]
name = "symptom-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-ollama>=1.1.0",
    "python-dotenv>=1.2.2",
    "reflex>=0.8.28.post1",
    "supabase>=2.28.3",
]
